# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTC App Api",
        "description": "MTC Application OpenApi Documentation",
        "contact": {
            "email": "aminetemadifar.w@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "localhost"
        }
    ],
    "paths": {
...
        "/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get all branches",
                "description": "Retrieve a list of all branches, optionally filtered by rank",
                "operationId": "843abaab6771f4896444bcb485b40646",
                "parameters": [
                    {
                        "name": "rank_id",
                        "in": "query",
                        "description": "Filter branches by rank ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of branches retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BranchResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "description": "Create a new branch with the provided data",
                "operationId": "485042f612d9f6379874055d5c687474",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "rank_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Branch"
                                    },
                                    "rank_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "color_name": {
                                        "type": "string",
                                        "example": "Blue"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BranchResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/branches/{branch}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get a specific branch",
                "description": "Retrieve details of a specific branch",
                "operationId": "85e093d89f51364a9a020fddd9288c5d",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BranchResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Branch not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Update a branch",
                "description": "Update the details of a specific branch",
                "operationId": "fa25c842dbb967daadae2af26242b56a",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Branch Name"
                                    },
                                    "rank_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "color_name": {
                                        "type": "string",
                                        "example": "Red"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/new-image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BranchResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Branch not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Delete a branch",
                "description": "Delete a specific branch",
                "operationId": "da80c3d8211bf60ae721c20978def14e",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Branch deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Branch not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BranchResource": {
                "title": "Branch Resource",
                "description": "Branch resource schema",
                "properties": {
                    "id": {
                        "description": "Branch code",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Branch name",
                        "type": "string",
                        "example": "Main Branch"
                    },
                    "rank_id": {
                        "$ref": "#/components/schemas/RankResource"
                    },
                    "color_name": {
                        "description": "Branch color name",
                        "type": "string",
                        "example": "Blue"
                    },
                    "group_id": {
                        "description": "Group code",
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "description": "Education chart image URL",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "status": {
                        "description": "Branch status",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "RankResource": {
                "title": "Rank Resource",
                "description": "Rank resource schema",
                "properties": {
                    "id": {
                        "description": "Rank code",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Rank name",
                        "type": "string",
                        "example": "Beginner"
                    },
                    "status": {
                        "description": "Rank status",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}